requirements.txt
mysql-connector-python==9.6.0
numpy==2.4.1
pandas==3.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
six==1.17.0
